warning avoided

+++ ql/Pricers/fddividendeuropeanoption.cpp
@@ -32,12 +32,12 @@ namespace QuantLib {
 
         for (Size j = 0; j < dividends.size(); j++) {
             QL_REQUIRE(exdivdates[j] >= 0, "The "+
-                       IntegerFormatter::toOrdinal(j) +
+                       IntegerFormatter::toOrdinal((unsigned long)(j)) +
                        " dividend time is negative"    + "(" +
                        DoubleFormatter::toString(exdivdates[j]) + ")");
 
             QL_REQUIRE(exdivdates[j] <= residualTime,"The " +
-                       IntegerFormatter::toOrdinal(j) +
+                       IntegerFormatter::toOrdinal((unsigned long)(j)) +
                        " dividend time is greater than residual time" + "(" +
                        DoubleFormatter::toString(exdivdates[j]) + ">" +
                        DoubleFormatter::toString(residualTime)    + ")");
